Feign client

+++ hr-payroll/src/main/java/com/devsuperior/hrpayroll/resources/PaymentResource.java
@@ -18,7 +18,7 @@ public class PaymentResource {
 	PaymentService service;
 	
 	@GetMapping(value = "/{workerId}/days/{days}")
-	public ResponseEntity<Payment> getPayment(@PathVariable Integer workerId, @PathVariable Integer days){
+	public ResponseEntity<Payment> getPayment(@PathVariable Long workerId, @PathVariable Integer days){
 		Payment obj = service.getPayment(workerId, days);
 		return ResponseEntity.ok(obj);
 	}
